% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacentWard.R
\name{adjacentWard}
\alias{adjacentWard}
\title{Adjacent Ward Cluster Algorithm.}
\usage{
adjacentWard(somobj, toroidal = FALSE)
}
\arguments{
\item{somobj}{an object of class "kohonen"}

\item{toroidal}{if TRUE, the edges of the map are joined. Note that in a hexagonal toroidal map, the number of rows must be even.}
}
\value{
List with cluster membership for different number of clusters.
}
\description{
Use the Adjacent Ward Cluster Algorithm to prevent discontiguous cluster 
segments. The implemented algorithm is described in further details in 
\url{https://www.viscovery.net/download/public/The-SOM-Ward-cluster-algorithm.pdf}.
}
\examples{
som_model <- som(as.matrix(iris[,1:4]), 
                 grid=somgrid(5,5,"hexagonal"))
adjacentWard(som_model, toroidal = FALSE)
}
