% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFunction_Factor2Dummies.R
\name{contr.onehot}
\alias{contr.onehot}
\title{Factors into dummies variables}
\usage{
contr.onehot(n, contrasts, sparse = FALSE)
}
\arguments{
\item{df}{- dataframe}
}
\value{
data frame without any factor variable (each original factor is replaced by a 
set of dummy variables.
}
\description{
Convert factor variable into multiple dummy variables
}
\examples{
\\code{RFunction_Factor2Dummies(iris)} - replaces factor (\\emph{Species}) into 3 dummies 
variables: \\emph{Speciessetosa, Speciesversicolor, Speciesvirginica}

}
