% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonalPlotEval.R
\name{hexagonalPlotEval}
\alias{hexagonalPlotEval}
\title{Plot one of many hexagonal evaluation plots.}
\usage{
hexagonalPlotEval(somobj, type, feature, ...)
}
\arguments{
\item{somobj}{an object of class "kohonen" or "somgrid".}

\item{type}{choose an evaluation plot. Possible values are "counts", "quality", "dist.neighbours" and "heatmap".}

\item{feature}{if \code{type = "heatmap"} the variable for the heatmap has to be defined.}
}
\value{
Plot a evaluation hexagonal plot.
}
\description{
hexagonalPlotEval allows to plot different evaluation plots in a hexagonal
layout.
}
