% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNode.R
\name{nearestNode}
\alias{nearestNode}
\title{Identify nearest node to click location.}
\usage{
nearestNode(somobj, click)
}
\arguments{
\item{somobj}{an object of class "kohonen" or "somgrid"}

\item{click}{x/y location of the user click}
}
\value{
Node number with smallest distance to clicked location.
}
\description{
After clicking into the interactive visualisation identify the nearest node 
based on euclidean distance.
}
\examples{
som_model <- som(as.matrix(iris[,1:4]), 
grid=somgrid(5,5,"hexagonal"))
nearestNode(somobj = som_model, click = c(2.5,1))

nearestNode(somobj = somgrid(5,5,"hexagonal"), click = c(2.5,1))
}
