% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Data winsorization}
\usage{
winsorize(x, minval = NULL, maxval = NULL, probs = c(0.05, 0.95))
}
\arguments{
\item{x}{a numeric vector to be winsorized.}

\item{minval}{the lower border, all values being lower than this will be replaced by this value.}

\item{maxval}{the upper border, all values being larger than this will be replaced by this value.}

\item{probs}{if minval or maxval is NULL use the given probabilities of quantile}
}
\value{
A vector of the same length as the original vector containing the winsorized data.
}
\description{
Make the data robust with the help of winsorization, i.e., by shrinking outlying observations to the border of the main part of the data.
}
\examples{
set.seed(42)
data <- rnorm(50)
summary(data)
summary(winsorize(data,probs=c(0.05, 0.95)))
summary(winsorize(data,probs=c(0.05, 1)))
}
